Evaluate paths instead of using their expression.
Prior to this commit, when an LDflex mutation would receive a path,
it would use that path's expression to describe the data in the query.
E.g., `data.user.friends.remove(data.user.enemies)` would take
the `enemies` expression and put the expression in the path.
However, this would not work if the path came from another object,
e.g., `data.user.friends.remove(data.other.enemies)`.

With this commit, paths passed to mutations are now evaluated,
and the resulting values are used to execute the operation.
E.g., `data.user.enemies` yields a list of terms,
and those terms are passed to the `remove` function.